Added manifest builder script
