+++ manifest_backup.json
@@ -37,15 +37,11 @@
         "scripts/angular-animate.min.js",
         "scripts/bootstrap-tpls.min.js",
         "js/init.js",
-        "js/controllers/all_data_controller.js",
-        "js/controllers/base_controller.js",
-        "js/directives/main_extension_dir.js",
-        "js/directives/text_collapse_dir.js",
-        "js/directives/toggle_button_dir.js",
-        "js/directives/toggle_extension_dir.js",
-        "js/directives/update_model_dir.js",
-        "js/route_config.js",
-        "js/services/data_service.js"
+        "js/routeConfig.js",
+        "js/allDataController.js",
+        "js/baseController.js",
+        "js/services.js",
+        "js/directives.js"
       ]
     },
     {
