+++ package.json
@@ -20,9 +20,6 @@
     "gulp": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.8.4",
-    "@babel/node": "^7.8.4",
-    "@babel/preset-env": "^7.8.4"
+    "glob": "^7.1.6"
   }
 }
