--- setup_scripts/manifest_builder.js
@@ -1,3 +1,0 @@
-import { a } from './test.js';
-
-console.log(a);
