--- setup_scripts/test.js
@@ -1,1 +1,0 @@
-export const a = 'TEST EXPORT';
